Fix #4090: clear the CI clippy gate workspace-wide
#4090's own scope named 7 sites (4 neg_cmp_op_on_partial_ord in
byroredux-core under CI's exact `cargo clippy --workspace -- -D
warnings`, plus 3 more under `--all-targets` in byroredux-renderer).
Fixing exactly those left the actual CI command still red: two days of
intervening commits had accumulated 26 more errors across
byroredux-core, byroredux-bsa, byroredux-nif, byroredux-plugin and
byroredux-renderer, and a further 33 in the byroredux binary itself —
none named by the issue, all newly surfaced once the named ones stopped
masking them. Fixed the lot; `cargo clippy --workspace -- -D warnings`
(CI's exact command) and `cargo clippy -p byroredux-renderer
--all-targets` are both clean.

The neg_cmp_op_on_partial_ord fixes (#4090's original 4-site scope):

- animation/player.rs, animation/stack.rs (×3), bsa/uvd.rs: rewrote
  `!(x >= y)` / `!(x < y)` degenerate-input guards as `x.is_nan() || x
  {<,>=} y`, equivalent for a non-NaN RHS (uvd.rs's RHS is also a
  runtime value, so both sides get an explicit is_nan() check) but
  legible as the deliberate NaN guard each already was, not an
  accidental double-negative.

- byroredux/src/systems/weather.rs: the four `.max(lo).min(hi)` clamps
  on `rain`/`snow`/`wind`/`sun` are kept in that form rather than
  clippy's suggested `.clamp(lo, hi)` — the two are NOT equivalent on a
  NaN input (`f32::clamp` returns NaN; the max-then-min chain returns
  `lo`), none of the four is guarded by an upstream `is_finite()` check
  the way `dt` two lines above is, and `WeatherSurfaceState` persists
  across frames, so a NaN let through here would latch in for the
  entity's life. Silenced with `#[allow(clippy::manual_clamp)]` and a
  comment rather than guessed into a behavior change; `dt` itself,
  which *is* is_finite()-guarded first, took the suggested rewrite.

Two genuine doc-attachment bugs found and fixed while chasing the
lints, not just silenced:

- crates/renderer/src/vulkan/ssao.rs: `initialize_ao_images`'s
  `# Safety` section was attached to the wrong item entirely — sitting
  above `ao_image_view` (a safe fn) with no doc at all on the actual
  `unsafe fn` below it. Moved to the right function.

- crates/renderer/src/mesh.rs: a doc block describing
  `GeometryRebuildInProgress` (its vertices_copied/indices_copied
  fields, the #3298/#3443 threshold history) was orphaned in mesh.rs
  documenting nothing — the struct it describes moved into
  mesh/geometry_ssbo.rs at some point and left this copy behind with no
  live subject. Relocated to the struct, which had no doc of its own.
  water_caustic.rs had the same shape but for a struct that had since
  collapsed into a type alias (`type Slot = GpuImage`) — that one was
  genuinely dead (a 3-field description of a shape that no longer
  exists) and was deleted rather than relocated, since the adjacent
  doc block already covers the current state including the collapse.

Everything else is mechanical: doc_lazy_continuation (a `+` at the
start of a doc line reads as an accidental markdown list bullet —
reworded to "and" in four places rather than indenting to preserve a
list that was never intended), unnecessary_cast, needless_borrow,
needless_option_as_deref, needless_range_loop, identity_op,
derivable_impls, unnecessary_map_or, large_enum_variant (boxed
InteriorCellApplyProgress::Pending's ~6KB payload; the one destructure
site derefs the box, no field-type changes elsewhere), and
too_many_arguments / type_complexity (allowed with a reason, or a named
type alias, matching this codebase's existing house style — 60
pre-existing too_many_arguments allows).

Verified: cargo clippy --workspace -- -D warnings clean; cargo clippy
-p byroredux-renderer --all-targets clean; cargo check --workspace
clean; _audit-validate.sh exit 0; full test suites green with no count
regressions — byroredux-core 755, byroredux-bsa 92, byroredux-nif 1270,
byroredux-renderer 981, byroredux --bin byroredux 2051, all 0 failed.

+++ byroredux/src/cell_loader/spawn/mesh_instance.rs
@@ -723,7 +723,7 @@ fn resolve_shared_geometry(
         return;
     };
     for &index in shared {
-        let representative = pc.geometry_dedup[index] as u32;
+        let representative = pc.geometry_dedup[index];
         match ctx.mesh_registry.acquire_cached(key, representative) {
             // `fresh_for_rt: false` — the representative already queued its
             // BLAS build this placement; a second request for the same mesh
